+++ src/handler/openGekkoUi.handler.ts
@@ -1,0 +1,6 @@
+import * as vscode from 'vscode';
+
+export default () => {
+    const gekkoAddress: string = vscode.workspace.getConfiguration().get('gekko.address');
+    vscode.env.openExternal(vscode.Uri.parse(gekkoAddress));
+};